% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keggDataREST.R
\name{KEGGpathway2Graph}
\alias{KEGGpathway2Graph}
\title{Modified version of the same function from KEGGgraph}
\usage{
KEGGpathway2Graph(pathway, genesOnly = TRUE, expandGenes = TRUE)
}
\description{
Modified version of the same function from KEGGgraph
}
\keyword{internal}

