% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{pDisPathway-class}
\alias{pDisPathway-class}
\title{Class that encodes the result of pDis analysis for a single pathway}
\description{
Class that encodes the result of pDis analysis for a single pathway
}
\section{Slots}{


\describe{
   \item{\code{map}:}{an object of type graph (e.g., \code{\link{graphNEL}}).}
   \item{\code{input}:}{named vector of fold changes for genes on this pathway. The names of the genes are the orignal IDS used in the analysis}
   \item{\code{ref}:}{vector of reference IDs on this pathway}
   \item{\code{boot}:}{an object of class \code{boot} encoding the bootstrap information.}
   \item{\code{pDis}:}{the gene primary dis-regulation for all genes on the pathway, as computed by primary dis-regulation.}
   \item{\code{asGS}:}{pathway was considered as gene set}
}
}
\author{
Calin Voichita, Sahar Ansari and Sorin Draghici
}
\seealso{
\code{\link{pDis}}, \code{\link{pDisRes-class}}
}

