% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{pePathway-class}
\alias{pePathway-class}
\title{Class that encodes the result of Pathway-Express for a single pathway}
\description{
Class that encodes the result of Pathway-Express for a single pathway
}
\section{Slots}{


\describe{
   \item{\code{map}:}{an object of type graph (e.g., \code{\link{graphNEL}}).}
   \item{\code{input}:}{named vector of fold changes for genes on this pathway. The names of the genes are the orignal IDS used in the analysis}
   \item{\code{ref}:}{vector of reference IDs on this pathway}
   \item{\code{boot}:}{an object of class \code{boot} encoding the bootstrap information.}
   \item{\code{Pert}:}{the gene perturbation factors for all genes on the pathway, as computed by Pathway-Express.}
   \item{\code{Acc}:}{the gene accumulations for all genes on the pathway, as computed by Pathway-Express.}
   \item{\code{asGS}:}{pathway was considered as gene set}
}
}
\author{
Calin Voichita and Sorin Draghici
}
\seealso{
\code{\link{pe}}, \code{\link{peRes-class}}
}

