% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keggDataREST.R
\name{subGraphByNodeType}
\alias{subGraphByNodeType}
\title{Modified version of the same function from KEGGgraph}
\usage{
subGraphByNodeType(graph, type = "gene")
}
\description{
Modified version of the same function from KEGGgraph
}
\keyword{internal}

