% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary-methods.R
\docType{methods}
\name{Summary,pDisRes-method}
\alias{Summary,pDisRes-method}
\title{Summarize the results of a Pathway-Express analysis}
\usage{
\S4method{Summary}{pDisRes}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{Primary dis-regulation analysis result object obtained using \code{\link{pDis}}}

\item{...}{see \code{\link{summary.pDisRes}}}

\item{na.rm}{ignored}
}
\description{
Summarize the results of a Pathway-Express analysis
}

