% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary-methods.R
\docType{methods}
\name{Summary,peRes-method}
\alias{Summary,peRes-method}
\title{Summarize the results of a Pathway-Express analysis}
\usage{
\S4method{Summary}{peRes}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{Pathway-Express analysis result object obtained using \code{\link{pe}}}

\item{...}{see \code{\link{summary.peRes}}}

\item{na.rm}{ignored}
}
\description{
Summarize the results of a Pathway-Express analysis
}

