% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute.normalInv}
\alias{compute.normalInv}
\title{Combine independent p-values using the normal inversion method}
\usage{
compute.normalInv(p, eps = 1e-06)
}
\arguments{
\item{p}{a vector of independent p-values}

\item{eps}{the minimal p-value considered (all p-values smaller will be set to this value)}
}
\value{
the combined p-value
}
\description{
Combine independent p-values using the normal inversion method
}
\examples{

p <- c(.1, .01)
compute.normalInv(p)

}
\author{
Calin Voichita and Sorin Draghici
}
\references{
Tarca AL., Draghici S., Romero R.: "A Mmore Specific Method To Combine Perturbation and Over-representation Evidence in Pathway Analysis", PSB 2010 poster.
}
\seealso{
\code{\link{pe}},\code{\link{compute.fisher}}
}

